dev-requirements.txt
pyspark>=3.3.0
delta-spark>=2.2.0
pandas<2.0.0
numpy<1.24
pyarrow
pytest
